quick save

+++ codeforlife/user/models/otp_bypass_token.py
@@ -12,27 +12,53 @@ from django.contrib.auth.hashers import check_password, make_password
 from django.core.exceptions import ValidationError
 from django.core.validators import MinLengthValidator
 from django.db import models
+from django.utils.translation import gettext_lazy as _
+from django_stubs_ext.db.models import TypedModelMeta
 
 from ...models import AbstractModel
-from . import user
+from . import user as _user
 
 
 class OtpBypassToken(AbstractModel):
+    """
+    A one-time-use token that a user can use to bypass their OTP auth factor.
+    Each user has a limited number of OTP-bypass tokens.
+    """
+
     max_count = 10
     max_count_validation_error = ValidationError(
         f"Exceeded max count of {max_count}"
     )
 
+    # pylint: disable-next=missing-class-docstring
     class Manager(models.Manager["OtpBypassToken"]):
-        def create(self, token: str, **kwargs):
+        def create(self, token: str, **kwargs):  # type: ignore[override]
+            """Create an OTP-bypass token.
+
+            Args:
+                token: The token value to be hashed.
+
+            Returns:
+                A OtpBypassToken instance.
+            """
+
             return super().create(token=make_password(token), **kwargs)
 
-        def bulk_create(
+        def bulk_create(  # type: ignore[override]
             self,
             otp_bypass_tokens: t.List["OtpBypassToken"],
             *args,
             **kwargs,
         ):
+            """Bulk create OTP-bypass tokens.
+
+            Args:
+                otp_bypass_tokens: The token values to be hashed.
+
+            Returns:
+                Many OtpBypassToken instances.
+            """
+
             def key(otp_bypass_token: OtpBypassToken):
                 return otp_bypass_token.user.id
 
@@ -40,7 +66,10 @@ class OtpBypassToken(AbstractModel):
             for user_id, group in groupby(otp_bypass_tokens, key=key):
                 if (
                     len(list(group))
-                    + OtpBypassToken.objects.filter(user_id=user_id).count()
+                    + OtpBypassToken.objects.filter(
+                        user_id=user_id,
+                        delete_after__isnull=True,
+                    ).count()
                     > OtpBypassToken.max_count
                 ):
                     raise OtpBypassToken.max_count_validation_error
@@ -50,9 +79,11 @@ class OtpBypassToken(AbstractModel):
 
             return super().bulk_create(otp_bypass_tokens, *args, **kwargs)
 
-    objects: Manager = Manager()
+    objects: Manager = Manager.from_queryset(  # type: ignore[misc]
+        AbstractModel.QuerySet
+    )()  # type: ignore[assignment]
 
-    user: "user.User" = models.ForeignKey(
+    user: "_user.User" = models.ForeignKey(  # type: ignore[assignment]
         "user.User",
         related_name="otp_bypass_tokens",
         on_delete=models.CASCADE,
@@ -63,13 +94,18 @@ class OtpBypassToken(AbstractModel):
         validators=[MinLengthValidator(8)],
     )
 
-    class Meta:
+    class Meta(TypedModelMeta):
+        verbose_name = _("OTP bypass token")
+        verbose_name_plural = _("OTP bypass tokens")
         unique_together = ["user", "token"]
 
     def save(self, *args, **kwargs):
         if self.id is None:
             if (
-                OtpBypassToken.objects.filter(user=self.user).count()
+                OtpBypassToken.objects.filter(
+                    user=self.user,
+                    delete_after__isnull=True,
+                ).count()
                 >= OtpBypassToken.max_count
             ):
                 raise OtpBypassToken.max_count_validation_error
@@ -77,7 +113,16 @@ class OtpBypassToken(AbstractModel):
         return super().save(*args, **kwargs)
 
     def check_token(self, token: str):
-        if check_password(token, self.token):
+        """Check if the token matches.
+
+        Args:
+            token: Token to check.
+
+        Returns:
+            A boolean designating if the token is matches.
+        """
+
+        if not self.delete_after and check_password(token, self.token):
             self.delete()
             return True
         return False
